refactor: move some functions to files
