+++ src/index.js
@@ -2,6 +2,8 @@ const readline = require("node:readline");
 const { stdin: input, stdout: output } = require("node:process");
 const Car = require("./Car");
 const AttemptCount = require("./AttemptCount");
+const getRandomIntInclusive = require("./getRandomIntInclusive");
+const findWinners = require("./findWinners");
 
 const rl = readline.createInterface({ input, output });
 
@@ -23,7 +25,8 @@ function askAttemptCount(cars) {
   rl.question("시도할 횟수는 몇 회인가요?\n", (answer) => {
     const attemptCount = new AttemptCount(answer);
     race(cars, attemptCount);
-    findWinners(cars);
+    const winners = findWinners(cars);
+    console.log(`최종 우승자: ${winners.map((it) => it.name)}`);
     rl.close();
   });
 }
@@ -39,27 +42,6 @@ function race(cars, attemptCount) {
   }
 }
 
-// https://developer.mozilla.org/ko/docs/Web/JavaScript/Reference/Global_Objects/Math/random
-function getRandomIntInclusive(min, max) {
-  min = Math.ceil(min);
-  max = Math.floor(max);
-  return Math.floor(Math.random() * (max - min + 1)) + min; //최댓값도 포함, 최솟값도 포함
-}
-
 function renderCar(car) {
   return `${car.name} : ${"-".repeat(car.position)}`;
-}
-
-function findWinners(cars) {
-  const maximumPosition = maxByPosition(cars);
-  const winners = filterByPosition(cars, maximumPosition);
-  console.log(`최종 우승자: ${winners.map((it) => it.name)}`);
-}
-
-function maxByPosition(cars) {
-  return Math.max(...cars.map((it) => it.position));
-}
-
-function filterByPosition(cars, position) {
-  return cars.filter((it) => it.position === position);
 }
